requirements.txt
opencv-python-headless
streamlit
pillow
numpy
plotly
streamlit-webrtc
av
ultralytics